Combined tests into one file
