--- src/qa/test/QuestionList.test.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import QuestionList from '../QuestionList.jsx';
-import { questions } from '../example-data/questionsData.js';
